nvim config, changes in readme and script installer

+++ install.sh
@@ -9,6 +9,7 @@ declare -A LINKS=(
   [bash_profile]="$HOME/.bash_profile"
   [xinitrc]="$HOME/.xinitrc"
   [rofi_config]="$HOME/.config/rofi/config.rasi"
+  [nvim_config]="$HOME/.config/nvim"
 )
 
 #funcion de backup para enlaces simbolicos seguros
@@ -16,24 +17,26 @@ backup() {
   src="$1"
   dst="$2"
 
-  [ -e "$dst" ] && mv "$dst" "$dst.bak"
+  if [ ! -e "$src" ]; then
+    echo "⚠️  No existe el origen: $src"
+    return
+  fi
+
+  mkdir -p "$(dirname "$dst")"
+
+  if [ -e "$dst" ] || [ -L "$dst" ]; then
+    mv "$dst" "$dst.bak"
+  fi
+
   ln -s "$src" "$dst"
 }
-
-echo "Creacion de enlaces simbolicos..."
-
 
 for file in "${!LINKS[@]}"; do
   echo "Creando enlace simbolico para .$file..."
   backup "$DOTFILES_DIR/$file" "${LINKS[$file]}"
 done
 
-echo "instalacion de nvim y NvChad..."
-
 sudo pacman -S --noconfirm nvim
-git clone https://github.com/NvChad/starter ~/.config/nvim
-
-echo "Para completar la instalacion de NvChad abrir nvim"
 
 #volver al home
 cd ~/
